Refactor folder structure: moved signup & create-post pages
