--- frontend/src/app/create-post/page.tsx
@@ -1,190 +1,0 @@
-"use client";
-
-import { useState } from "react";
-import { useDropzone } from "react-dropzone";
-import { useRouter } from "next/navigation";
-import { Button } from "@/components/ui/button";
-import { v4 as uuidv4 } from "uuid";
-import { createPost } from "../api/create-post_page";
-
-export default function CreatePostPage() {
-  const [tab, setTab] = useState<"text" | "media">("text");
-  const [formData, setFormData] = useState({
-    title: "",
-    body_md: "",
-    url: "",
-    category_id: "",
-    files: [] as File[]
-  });
-
-  const [loading, setLoading] = useState(false);
-  const [message, setMessage] = useState("");
-
-  const router = useRouter()
-
-  const onDrop = (acceptedFiles: File[]) => {
-    setFormData((prev) => ({
-      ...prev,
-      files: [...prev.files, ...acceptedFiles],
-    }));
-  };
-
-  const { getRootProps, getInputProps, isDragActive } = useDropzone({
-    onDrop,
-    accept: {
-      "image/*": [],
-      "video/*": [],
-    },
-  });  
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-
-    const postId = uuidv4();
-    const postUrl = `http://post/${postId}`;
-
-    try {
-      const res = await createPost({ ...formData, url: postUrl});
-      router.push("/dashboard")
-      setMessage("Post created! ID: " + res.id);
-    } catch (err: any) {
-      console.log("Error: " + err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="max-w-3xl mx-auto p-4 bg-white rounded-md shadow">
-      <h1 className="text-2xl font-bold mb-4">Create Post</h1>
-      <div className="flex gap-4 border-b mb-4 bg-white">
-        <button
-          type="button"
-          onClick={() => setTab("text")}
-          className={`py-2 px-4 transition-colors duration-200 ${
-            tab === "text"
-              ? "border-b-4 border-blue-500 font-semibold"
-              : "font-semibold hover:bg-gray-100"
-          }`}
-        >
-          Text
-        </button>
-
-        <button
-          type="button"
-          onClick={() => setTab("media")}
-          className={`py-2 px-4 transition-colors duration-200 ${
-            tab === "media"
-              ? "border-b-4 border-blue-500 font-semibold"
-              : "font-semibold hover:bg-gray-100"
-          }`}
-        >
-          Images & Video
-        </button>
-      </div>
-      <form onSubmit={handleSubmit} className="flex flex-col gap-4">
-        <input
-          type="text"
-          name="title"
-          placeholder="Title"
-          value={formData.title}
-          onChange={(e) => setFormData({ ...formData, title: e.target.value })}
-          className="border rounded-xl p-3"
-          required
-        />
-
-        <select
-          className="border rounded-xl p-3"
-          value={formData.category_id}
-          onChange={(e) => setFormData({ ...formData, category_id: e.target.value })}
-        >
-          <option value="">Select Category</option>
-          <option value="64D7423E-F36B-1410-84C9-00F2EA0D0522">Game</option>
-        </select>
-
-        {tab === "text" && (
-          <textarea
-            placeholder="Body text (optional)"
-            className="border p-3 rounded-xl h-40"
-            value={formData.body_md}
-            onChange={(e) => setFormData({ ...formData, body_md: e.target.value })}
-          />
-        )}
-
-        {tab === "media" && (
-          <div className="flex flex-col gap-4">
-            <div
-              {...getRootProps()}
-              className={`border-2 border-dashed rounded-xl p-6 text-center cursor-pointer min-h-[200px] flex items-center justify-center ${
-                isDragActive ? "bg-blue-50 border-blue-400" : "border-gray-300"
-              }`}
-            >
-              <input {...getInputProps()} />
-              {formData.files.length === 0 ? (
-                <p>
-                  {isDragActive
-                    ? "Drop files here..."
-                    : "Drag & drop images/videos here, or click to select"}
-                </p>
-              ) : (
-                <div className="grid grid-cols-3 gap-4 w-full">
-                  {formData.files.map((file, i) => (
-                    <div key={i} className="relative group">
-                      {file.type.startsWith("image/") ? (
-                        <img
-                          src={URL.createObjectURL(file)}
-                          alt={file.name}
-                          className="h-32 w-full object-cover rounded"
-                        />
-                      ) : (
-                        <video
-                          controls
-                          className="h-32 w-full object-cover rounded"
-                          src={URL.createObjectURL(file)}
-                        />
-                      )}
-
-                      <button
-                        type="button"
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          setFormData((prev) => ({
-                            ...prev,
-                            files: prev.files.filter((_, idx) => idx !== i),
-                          }));
-                        }}
-                        className="absolute top-1 right-1 bg-black bg-opacity-50 text-white rounded-full pl-2 pr-2 pt-1 pb-1 text-xs opacity-0 group-hover:opacity-100 transition"
-                      >
-                        ✕
-                      </button>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </div>
-            <textarea
-              placeholder="Body text"
-              className="border p-3 rounded-xl h-40"
-              value={formData.body_md}
-              onChange={(e) => setFormData({ ...formData, body_md: e.target.value })}
-            />
-          </div>
-        )}
-
-        <div className="flex justify-end gap-2">
-          <Button
-            type="submit"
-            disabled={loading}
-            className="bg-emerald-800 rounded-full hover:bg-emerald-900"
-          >
-            Post
-          </Button>
-        </div>
-      </form>
-
-      {message && <p className="mt-4">{message}</p>}
-    </div>
-  );
-}
-
